split out into separate files.
switched to logging module
